:truck: Move notifiers to own directory
