+++ source/notifier/AlbumDOMBasedNotifier.js
