+++ source/notifier/StickerAlbumNotifier.js
